Slime Patron: Refactor all acid stuff to spike stuff
