+++ src/main/resources/anniv7Resources/images/characters/SpikeSlime/Portrait2.png
