+++ src/main/resources/anniv7Resources/images/characters/SpikeSlime/Portrait3.png
