+++ src/main/resources/anniv7Resources/images/relics/SpikeSlime/HandfulOfSlime.png
