+++ src/main/resources/anniv7Resources/images/relics/SpikeSlime/HandfulOfSlimeOutline.png
