--- src/main/resources/anniv7Resources/localization/eng/AcidSlimeBPatron/Characterstrings.json
@@ -1,9 +1,0 @@
-{
-  "${ModID}:AcidSlimeBPatron": {
-    "NAMES": [
-      "Acid Slime B"
-    ],
-    "TEXT": [
-    ]
-  }
-}
