+++ src/main/resources/anniv7Resources/localization/eng/SpikeSlimePatron/Characterstrings.json
@@ -1,0 +1,9 @@
+{
+  "${ModID}:SpikeSlimePatron": {
+    "NAMES": [
+      "Spike Slime"
+    ],
+    "TEXT": [
+    ]
+  }
+}
