+++ src/main/resources/anniv7Resources/localization/eng/SpikeSlimePatron/Relicstrings.json
